server/pyproject.toml
[project]
name = "mcp-streamable-http-server"
version = "0.1.0"
description = "HTTP Streamable based example MCP Server"
requires-python = ">=3.12"
dependencies = [
    "httpx~=0.28.0",
    "mcp~=1.9.0",
    "python-dotenv~=1.1.0",
    "pyjwt~=2.10.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
